create visibility style and refactor
