+++ src/style/shared/visibility.css.ts
@@ -1,8 +1,8 @@
 import { styleVariants } from "@vanilla-extract/css";
 
 export const visibilityVariants = styleVariants({
-    visible: { visibility: "visible" },
-    hidden: { visibility: "hidden" },
+  visible: { visibility: "visible" },
+  hidden: { visibility: "hidden" },
 });
 
 export type Visibility = keyof typeof visibilityVariants | (string & {});
